feat: wire memory, verification and injection defence into live paths
Closes G3, G5, G6 and G9 from docs/DEFINITION_OF_DONE.md.

Three subsystems were built, documented and tested but had no caller, so two
advertised behaviours did not happen: submission proof was never verified, and
the AI memory recorded corrections, embedded and ranked them, then dropped
them. sanitize_external_text was specified in CONTRACTS 10b and never written
at all, leaving job descriptions -- attacker-controlled text -- going straight
into prompts.

The PII screen landed before the memory wiring on purpose. ReviewService
stores the human's literal answer to a form field, so a reviewer typing an SSN
creates a memory containing it, and redact_secrets is key-based and log-scoped
-- it would not catch a bare value in a memory body and does not run on prompts
at all. Wiring memory first would have created the leak and then closed it.

Memories go into the system prompt as style constraints only, never into the
facts block, so golden rule 7 stays authoritative. Verified adversarially: a
memory asserting 'I also worked at Google' produced no surviving bullet -- the
fabricated fact id was dropped, the rewrite reverted to the source fact, and
the headers were copied from the fact rather than the model.

Verification now runs after a real submit and maps all four outcomes, with
inconclusive going to NEEDS_REVIEW rather than assuming success. AutoFiller
still skips file inputs -- the confidence machinery has nothing to say about
one -- but the apply driver reconciles them itself, so an application can no
longer be submitted with no resume attached.

Smoke test now seeds its own account, posting and signal so the five apply
flows actually run, then deletes them and exits non-zero if anything survives:
85 passed, 0 skipped. ruff is clean; the E501s were rewrapped rather than
suppressed.

+++ app/ai/scoring.py
@@ -2052,7 +2052,7 @@ def _clamp_adjustment(value: Any) -> int:
     if isinstance(value, bool) or value is None:
         return 0
     try:
-        adjustment = int(round(float(value)))
+        adjustment = round(float(value))
     except (TypeError, ValueError):
         return 0
     return max(-MAX_LLM_ADJUSTMENT, min(MAX_LLM_ADJUSTMENT, adjustment))
